+++ src/main/clojure/com/github/clojure_repl/intellij/app_info.clj
@@ -1,0 +1,9 @@
+(ns com.github.clojure-repl.intellij.app-info
+  (:import
+   [com.intellij.openapi.application ApplicationInfo]
+   [com.intellij.openapi.util BuildNumber]))
+
+(defn at-least-version? [version]
+  (>= (.compareTo (.withoutProductCode (.getBuild (ApplicationInfo/getInstance)))
+                   (BuildNumber/fromString version))
+      0))