+++ src/main/clojure/com/github/clojure_repl/intellij/ui/hint.clj
@@ -1,6 +1,9 @@
 (ns com.github.clojure-repl.intellij.ui.hint
+  (:require
+   [com.github.clojure-repl.intellij.app-info :as app-info])
   (:import
-   [com.intellij.codeInsight.hint HintManager]
+   [com.intellij.codeInsight.hint
+    HintManager]
    [com.intellij.openapi.editor Editor]))
 
 (set! *warn-on-reflection* true)
@@ -15,10 +18,14 @@
       :or {prefix ""}}]
   (.showErrorHint (HintManager/getInstance) ^Editor editor (str prefix message) HintManager/RIGHT))
 
+(set! *warn-on-reflection* false)
 (defn show-success
-  [& {:keys [message editor prefix]
+  [& {:keys [message ^Editor editor prefix]
       :or {prefix ""}}]
-  (.showInformationHint (HintManager/getInstance) ^Editor editor (str prefix message) HintManager/RIGHT))
+  (if (app-info/at-least-version?  "233.9802.14")
+    (.showSuccessHint (HintManager/getInstance) ^Editor editor (str prefix message) HintManager/RIGHT)
+    (.showInformationHint (HintManager/getInstance) ^Editor editor (str prefix message) HintManager/RIGHT)))
+(set! *warn-on-reflection* true)
 
 (defn show-repl-info [& {:as args}]
   (show-info (assoc args :prefix "=> " :message (or (:message args) "nil"))))